Turn AsyncRuntime into singleton
